consolidate fixed binary to plain encoding
