+++ protos/format.proto
@@ -72,7 +72,6 @@ enum Encoding {
   PLAIN = 1;
   VAR_BINARY = 2;
   DICTIONARY = 3;
-  FIXED_SIZE_BINARY = 4;
 }
 
 /**
